JAXB tests split.
