+++ moxy/eclipselink.moxy.test/antbuild.xml
@@ -561,10 +561,16 @@
                 <syspropertyset refid="parser.properties"/>
                 <batchtest todir="${report.dir}/jaxb">
                     <fileset dir="${src.dir}">
-                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite.java"/>
-                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite2.java"/>
-                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite3.java"/>
-                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite4.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite1A.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite1B.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite1C.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite2A.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite2B.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite2C.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite3A.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite3B.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite4A.java"/>
+                        <include name="org/eclipse/persistence/testing/jaxb/JAXBTestSuite4B.java"/>
                         <include name="org/eclipse/persistence/testing/jaxb/JAXBBeanValidationTestSuite.java"
                                  if="validate.jar.exists"/>
                         <include name="org/eclipse/persistence/testing/jaxb/beanvalidation/BeanValidationHelperTestCase.java"
