+++ moxy/eclipselink.moxy.test/src/org/eclipse/persistence/testing/jaxb/JAXBTestSuite3B.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1998, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1998, 2022 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0 which is available at
@@ -14,11 +14,6 @@
 //     Oracle - initial API and implementation from Oracle TopLink
 package org.eclipse.persistence.testing.jaxb;
 
-import junit.framework.Test;
-import junit.framework.TestCase;
-import junit.framework.TestSuite;
-
-import org.eclipse.persistence.testing.jaxb.interfaces.xmltransient.InvalidTransientInterfaceTestCases;
 import org.eclipse.persistence.testing.jaxb.jaxbcontext.duplicateelem.ObjectFactoryTestCases;
 import org.eclipse.persistence.testing.jaxb.jaxbcontext.withjaxbindex.JAXBContextByClassArrayWithIndexTestCases;
 import org.eclipse.persistence.testing.jaxb.jaxbcontext.withjaxbindex.JAXBContextByClassArrayWithRefInBindingsTestCases;
@@ -33,8 +28,6 @@ import org.eclipse.persistence.testing.jaxb.prefixmapper.PrefixMapperTestCases;
 import org.eclipse.persistence.testing.jaxb.properties.PropertyTestCases;
 import org.eclipse.persistence.testing.jaxb.readonly.ReadAndWriteOnlyTestCases;
 import org.eclipse.persistence.testing.jaxb.security.xss.SecurityXSSTestCases;
-import org.eclipse.persistence.testing.jaxb.stax.XMLStreamReaderEndEventTestCases;
-import org.eclipse.persistence.testing.jaxb.stax.XMLStreamWriterDefaultNamespaceTestCases;
 import org.eclipse.persistence.testing.jaxb.unmapped.UnmappedElementsWarningTestCases;
 import org.eclipse.persistence.testing.jaxb.unmarshaller.DefaultValueTestCases;
 import org.eclipse.persistence.testing.jaxb.unmarshaller.JSONUnmarshalAutoDetectTestCases;
@@ -47,11 +40,11 @@ import org.eclipse.persistence.testing.jaxb.unmarshaller.space.UnmarshalWithSpac
 import org.eclipse.persistence.testing.jaxb.unmarshaller.validation.ValidationTestCases;
 import org.eclipse.persistence.testing.jaxb.uri.ChildURITestCases;
 import org.eclipse.persistence.testing.jaxb.uri.URITestCases;
+import org.eclipse.persistence.testing.jaxb.xmlAnyMixed.XmlAnyMixedTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlbindings.XMLBindingsPopulatedTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlbindings.XMLBindingsTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlbindings.XMLBindingsWithExternalMetadataTestCases;
 import org.eclipse.persistence.testing.jaxb.xmldecriptor.LazyInitTestCases;
-import org.eclipse.persistence.testing.jaxb.xmlAnyMixed.XmlAnyMixedTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlmixed.XmlMixedTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlpath.XmlPathToAttributeTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlpath.XmlPathToElementWithXmlAttributeTestCases;
@@ -60,60 +53,20 @@ import org.eclipse.persistence.testing.jaxb.xmlpath.XmlPathWithXmlAttributeTestC
 import org.eclipse.persistence.testing.jaxb.xmlschema.XMLSchemaModelTestCases;
 import org.eclipse.persistence.testing.jaxb.xmlvirtualaccessmethods.proporder.PropOrderTestCases;
 
-public class JAXBTestSuite3 extends TestCase {
+import junit.framework.Test;
+import junit.framework.TestCase;
+import junit.framework.TestSuite;
+
+public class JAXBTestSuite3B extends TestCase {
 
     public static Test suite() {
-        TestSuite suite = new TestSuite("JAXB20 Compiler Test Suite3");
-
-        suite.addTest(org.eclipse.persistence.testing.jaxb.singleobject.JAXBSingleObjectTestSuite.suite());
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.classloader.DifferentClassLoaderTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.classloader.InnerClassTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.classloader.XmlElementsEnumTestCases.class);
-
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.helper.JAXBHelperTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.any.AnyWithJAXBElementTestCases.class);
-
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.interfaces.InterfaceTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.interfaces.choice.InterfaceChoiceTestCases.class);
-        suite.addTestSuite(InvalidTransientInterfaceTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.inheritance.interfaces.InterfacesTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.MarshalSchemaValidationTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.NoSchemaRefTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.UnmarshalSchemaValidationTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.UnmarshallerNullTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.NoSchemaRefTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.HandleListenerExceptionsTestCases.class);
-        suite.addTestSuite(XMLStreamWriterDefaultNamespaceTestCases.class);
-        suite.addTestSuite(XMLStreamReaderEndEventTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.qname.QNameTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.qname.defaultnamespace.QNameTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.locator.AnyTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.locator.AnyCollectionTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.locator.ElementTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlmarshaller.locator.ElementCollectionTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmltype.XmlTypeTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmltype.XmlTypeNameTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmltype.proporder.NonTransientTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmltype.proporder.TransientTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmltype.proporder.ExtraPropTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmltype.proporder.MissingPropTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlaccessortype.none.NoneTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlaccessortype.FieldAndPropertyTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlaccessortype.IgnoreInvalidNonPublicFieldTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlvirtualaccessmethods.XmlVirtualAccessMethodsTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlschematype.XmlSchemaTypeDateTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlschematype.XmlSchemaTypeDateEmptyTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlschematype.XmlSchemaTypeTwoDatesTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlschematype.NonNegativeIntegerSchemaTypeTestCases.class);
-        suite.addTestSuite(org.eclipse.persistence.testing.jaxb.xmlschematype.CharTestCases.class);
-
+        TestSuite suite = new TestSuite("JAXB20 Compiler Test Suite 3B");
         suite.addTestSuite(JAXBContextByClassArrayWithIndexTestCases.class);
         suite.addTestSuite(JAXBContextByClassArrayWithRefTestCases.class);
         suite.addTestSuite(JAXBContextByClassArrayWithRefInBindingsTestCases.class);
         suite.addTestSuite(JAXBContextByPackageWithIndexTestCases.class);
         suite.addTestSuite(ObjectFactoryTestCases.class);
         //suite.addTestSuite(MultipleXmlElementRefTestCases.class);
-
         suite.addTestSuite(PropOrderTestCases.class);
         suite.addTestSuite(LazyInitTestCases.class);
         suite.addTestSuite(XmlMixedTestCases.class);
@@ -147,12 +100,11 @@ public class JAXBTestSuite3 extends TestCase {
         suite.addTestSuite(PropertyTestCases.class);
         suite.addTestSuite(UnmappedElementsWarningTestCases.class);
         suite.addTestSuite(SecurityXSSTestCases.class);
-
         return suite;
     }
 
     public static void main(String[] args) {
-        String[] arguments = { "-c", "org.eclipse.persistence.testing.jaxb.JAXBTestSuite3" };
+        String[] arguments = { "-c", "org.eclipse.persistence.testing.jaxb.JAXBTestSuite3B" };
         // junit.swingui.TestRunner.main(arguments);
         // System.setProperty("useLogging", "true");
         junit.textui.TestRunner.main(arguments);
